ksrc, scripts: remove pre-2.6.30 kernel support
One of the promises of the 3.x architecture is to get rid of the
ancient kernel legacy, which is 1) a pain to maintain via hairy
wrappers, 2) a restriction on using recent kernel features to improve
the Xenomai core.

This patch cuts off pre-2.6.30 support for now, but it is very likely
that we will move this barrier up, depending on what non-major CPU
arch ports can be based on in a reasonable future (typically nios).