testsuite: remove irqbench
This test depends on having a high priority domain on top of Xenomai,
which won't be supported in the next major architecture.

The point is that the impact on performance of virtualizing the
interrupt masking for Xenomai - as for any non-heading domain - is too
large, so interpreting results wrt latency issues becomes problematic.

Besides, domain interposition for tracing/debugging purposes can now
be replaced by observation directly from within the Xenomai core or
via dedicated tools such as the I-pipe tracer.