copperplate/mercury: allow SUSP and PEND states to cumulate
SUSP and PEND(+DELAY) states must be cumulative,
i.e. threadobj_suspend() on a thread already blocked in syncobj_pend()
must hold it until threadobj_resume() is called, even if the
synchronization object is posted in the meantime.

This patch allows the notification signal to restart syncobj_pend()
once the handler returns, so that the situation looks like the
suspension has cumulated with the pending state.