copperplate/heapobj: fix tlsf-based heap locking
It has only recently be discovered that TLSF would not apply proper
locking when called for any of its *_ex() operations. Despite what
-DTLSF_USE_LOCKS=1 may suggest to the optimistic, this setting only
applies to the outer tlsf_*() entry points working on the main pool.

In short, TLSF leaves proper locking to the caller for using its
pool-oriented interface.

This patch introduces proper locking into the copperplate code
wrapping around TLSF *_ex() calls, adding a mutex into the heapobj
descriptor, which incidentally leads to having a superfluous lock in
each pool header on the TLSF-side, which is not for the main
one. Great.