cobalt/nucleus: remove support for priority coupling (aka RPI)
The former Xenomai architecture provided for priority coupling, a
mechanism which makes the Linux kernel - as a whole - inherit the
Xenomai-defined priority for a real-time thread running in user-space,
when it enters the secondary execution mode, e.g. to perform regular
Linux system calls.

The purpose, pros and cons of such feature has been discussed in
length over the years on the Xenomai mailing lists. A general
consensus among contributors emerged about the limitations of such
mechanism for properly tracking and propagating priority changes,
which cannot be lifted due to dual kernel design issues.

As a matter of fact, these limitations can be confusing to the end
user, and may have caused wrong application designs in some
occasions. Therefore we remove this support completely, moving back to
the principle of least suprise in dual kernel environments,
i.e. calling services of the non real-time kernel ALWAYS demotes a
real-time caller to the lowest, non real-time priority level.