cobalt: introduce wrapped/non-wrapped POSIX call annotation
__RT() and __STD() macros are introduced to enclose POSIX calls which
should go through the corresponding Cobalt real-time wrapper
(__wrap_*) or directly to the standard glibc routine (__real_*)
respectively.

At this chance, we sanitize the declarations of all routines we
overload in the Cobalt interface, providing prototypes for both the
wrapped and non-wrapped forms, all grouped in the new
include/cobalt/wrappers.h file.

The intent is to implement self-documenting and non-ambiguous code
using explicit annotation of POSIX calls, with respect to the actual
call we want depending on the underlying real-time core
(cobalt/mercury).

We make the Cobalt API use such annotation as appropriate, other
interfaces will follow.