copperplate: annotate wrapped/non-wrapped POSIX calls
We want non-POSIX APIs based on the Copperplate interface not to
require symbol wrapping via the linker's --Wl,--wrap magic. The latter
should only be required in presence of POSIX application code.

Rationale: the end user may want to use the standard
(i.e. non-wrapped) POSIX routines by default when developing over the
non-POSIX Xenomai APIs. Therefore we should not impose the linker
trick on him as a side-effect of depending on the Copperplate API
indirectly.

To this end, we annotate each POSIX call issued by the Copperplate
interface which may be shadowed by the Cobalt API, to specify whether
we want to call the wrapped version explicitly when running over the
Cobalt core, or whether we want the standard libc service to be
invoked.

We do this by enclosing the relevant POSIX calls within the __RT() or
__STD() annotations respectively. The final mapping will be as
follows:

core/call               __RT(foo)       __STD(foo)
--------------------------------------------------
cobalt                  __wrap_foo()    __real_foo()
mercury                 foo()           foo()

We introduce include/copperplate/wrappers.h to define the __RT() and
__STD() annotations conditionally, depending on the real-time core we
build Copperplate for.