copperplate/clockobj: introduce config switch for using CLOCK_MONOTONIC_RAW
We may run over vanilla kernels which do not support
CLOCK_MONOTONIC_RAW, albeit the glibc we built the Mercury core
against does advertise it.

This patch introduces the --enable-clock-monotonic-raw config swich to
make the copperplate interface use/ignore this feature at build time.
If unavailable, copperplate falls back to CLOCK_MONOTONIC. This is the
default setting for Mercury builds.

Note: The Cobalt core does support CLOCK_MONOTONIC_RAW
unconditionally, so the default is to enable it.