copperplate/threadobj: introduce wait structs
We need a conventional and convenient way to store per-thread private
information shared between a waiter-side service, and the
corresponding poster-side routine (e.g. such as wait_event and
signal_event).

Such data would usually define the wait context in a structure, to be
shared between both routines, so that a waiter can specify details
about what is being waited for to the poster, and the poster can peek
at the context of each waiter to determine whether the wait condition
is satisfied, waking up the thread if so.

To this end, we provide threadobj_alloc_wait(), threadobj_free_wait()
and threadobj_get_wait(), respectively to allocate, release and
retrieve a wait context. The address of a wait context is kept into
the control block of the thread who allocated it for waiting. A single
wait context is allowed per-thread.

In shared processing mode, the wait context is allocated by
threadobj_alloc_wait() from the shared heap, so that remote processes
can access it safely. Otherwise, it is allocated from the caller's
stack.

A typical usage would be something like this:

void wait_something(...)
{
	struct some_wait_context *p;

	...
	syncobj_lock(&sobj, &syns);
	p = threadobj_alloc_wait(struct some_wait_context);
	p->wait_for = 42;
	...
	ret = syncobj_pend(&sobj, &timespec, &syns);
	...
	threadobj_free_wait(p);
	syncobj_unlock(&sobj, &syns);
	...
}

void signal_something(int value)
{
	struct threadobj *thobj, *tmp;
	struct some_wait_context *p;

	...
	syncobj_lock(&sobj, &syns);
	...
	syncobj_for_each_waiter_safe(&sobj, thobj, tmp) {
		p = threadobj_get_wait(thobj);
		if (p->wait_for == value)
			syncobj_wakeup_waiter(&sobj, thobj);
	}
	...
	syncobj_unlock(&sobj, &syns);
	...
}

The wait context could also be used to pass back information from the
poster to the waiter upon wake up, the same way.

NOTE: this mechanism deprecates the wait_u struct still present in the
basic thread control block, which is scheduled for removal.